Clean ExecutionBatch to show that it's a service
This commit does nothing safe removing unused abstraction (ExecutionBatch trait),
unused methods, adapt tests in consequences, remove test about size of seq (no
more Rudder logic here), make private whatever can be.
With all that, it appears that ExecutionBatch is not a data structure but a
service (and a pure one!)
Next-step: making it an actual service, and make it depends upon
the type of compliance mode we are in.